requirements.txt
fastapi
uvicorn
langchain-community
langchain-core
openai
chromadb
pydantic
tiktoken
python-dotenv
PyJWT